+++ requests_cache/models/response.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from datetime import datetime, timedelta, timezone, UTC
+from datetime import datetime, timedelta
 from logging import getLogger
 from time import time
 from typing import TYPE_CHECKING, Dict, List, Optional, Union
@@ -11,7 +11,7 @@ from requests import PreparedRequest, Response
 from requests.cookies import RequestsCookieJar
 from requests.structures import CaseInsensitiveDict
 
-from ..policy import ExpirationTime, get_expiration_datetime
+from ..policy import ExpirationTime, get_expiration_datetime, utcnow
 from . import CachedHTTPResponse, CachedRequest, RichMixin
 
 if TYPE_CHECKING:
@@ -28,7 +28,7 @@ class BaseResponse(Response):
     provide type hints for extra cache-related attributes that are added to non-cached responses.
     """
 
-    created_at: datetime = field(factory=lambda: datetime.now(UTC))
+    created_at: datetime = field(factory=utcnow)
     expires: Optional[datetime] = field(default=None)
     cache_key: str = ''  # Not serialized; set by BaseCache.get_response()
     revalidated: bool = False  # Not serialized; set by CacheActions.update_revalidated_response()
@@ -54,7 +54,7 @@ class OriginalResponse(BaseResponse):
             # Add expires and cache_key only if the response was written to the cache
             response.expires = None if actions.skip_write else actions.expires  # type: ignore
             response.cache_key = None if actions.skip_write else actions.cache_key  # type: ignore
-            response.created_at = datetime.now(UTC)  # type: ignore
+            response.created_at = utcnow()  # type: ignore
         return response  # type: ignore
 
 
@@ -80,7 +80,7 @@ class CachedResponse(RichMixin, BaseResponse):
 
     def __attrs_post_init__(self):
         # Not using created_at field default due to possible bug on Windows with omit_if_default
-        self.created_at = self.created_at or datetime.now(UTC)
+        self.created_at = self.created_at or utcnow()
         # Re-initialize raw (urllib3) response after deserialization
         self.raw = self.raw or CachedHTTPResponse.from_cached_response(self)
 
@@ -130,7 +130,7 @@ class CachedResponse(RichMixin, BaseResponse):
         """Get time to expiration in seconds (rounded to the nearest second)"""
         if self.expires is None:
             return None
-        delta = self.expires - datetime.now(UTC)
+        delta = self.expires - utcnow()
         return round(delta.total_seconds())
 
     @property
@@ -146,7 +146,7 @@ class CachedResponse(RichMixin, BaseResponse):
     @property
     def is_expired(self) -> bool:
         """Determine if this cached response is expired"""
-        return self.expires is not None and datetime.now(UTC) >= self.expires
+        return self.expires is not None and utcnow() >= self.expires
 
     def is_older_than(self, older_than: ExpirationTime) -> bool:
         """Determine if this cached response is older than the given time"""