+++ HISTORY.md
@@ -4,6 +4,7 @@
 * ⚠️ Remove methods [deprecated in 1.0](#deprecations-1-0) from `CachedSession` and `BaseCache`
 * Optimize `SQLiteCache.delete()` when deleting a single key
 * Add `vacuum` parameter to `SQLiteCache.delete()` to optionally skip vacuuming after deletion (enabled by default to free up disk space)
+* Use timezone-aware UTC datetimes for all internal expiration values
 
 ## 1.1.0 (2023-06-30)